rename completeTaskActionModalData to taskActionModalData
